Add snake self collision and creating target
Added 2d array and vector to the game to implement the snake self
collision detection the michanism is to use the 2d bool array to
know if every tile is filled(with snake) or not.
the vector is used to prevent the positioning of target on the snake
it is filled with the indexes of empty tiles so I pick
random index from the vector and convert it to the x and y of target.
I also made some small code refactoring